BQ: move *QueryParameter and UDFResource classes to query. (#4156)
These classes are expected to be constructed by users, so moving them to
a public module.
