+++ bigquery/google/cloud/bigquery/__init__.py
@@ -26,10 +26,6 @@ The main concepts with this API are:
 from pkg_resources import get_distribution
 __version__ = get_distribution('google-cloud-bigquery').version
 
-from google.cloud.bigquery._helpers import ArrayQueryParameter
-from google.cloud.bigquery._helpers import ScalarQueryParameter
-from google.cloud.bigquery._helpers import StructQueryParameter
-from google.cloud.bigquery._helpers import UDFResource
 from google.cloud.bigquery._helpers import Row
 from google.cloud.bigquery.client import Client
 from google.cloud.bigquery.dataset import AccessEntry
@@ -39,6 +35,10 @@ from google.cloud.bigquery.job import CopyJobConfig
 from google.cloud.bigquery.job import ExtractJobConfig
 from google.cloud.bigquery.job import QueryJobConfig
 from google.cloud.bigquery.job import LoadJobConfig
+from google.cloud.bigquery.query import ArrayQueryParameter
+from google.cloud.bigquery.query import ScalarQueryParameter
+from google.cloud.bigquery.query import StructQueryParameter
+from google.cloud.bigquery.query import UDFResource
 from google.cloud.bigquery.schema import SchemaField
 from google.cloud.bigquery.table import Table
 from google.cloud.bigquery.table import TableReference
